✨ feat: implement emit_distress for critical anomaly alerts & add MetricAnalyzer
