+++ apps/server/app/core/helpers/emit_distress.py
@@ -1,0 +1,10 @@
+from app.utils.logging import create_logger
+
+logger = create_logger(__name__)
+
+
+def emit_distress():
+    """Mock method to emit distress signals when anomalies are detected"""
+    logger.critical("🚨 DISTRESS SIGNAL EMITTED - Critical anomaly detected!")
+    # In real implementation, this would trigger alerts, notifications, etc.
+    pass
