+++ apps/server/app/core/types/anomaly.py
@@ -1,17 +1,16 @@
 from datetime import datetime
 from enum import Enum
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Optional
 
 from pydantic import BaseModel
 
 
 class AnomalyType(Enum):
-    METRIC_SPIKE = "metric_spike"
-    ERROR_BURST = "error_burst"
-    PERFORMANCE_DEGRADATION = "performance_degradation"
-    SERVICE_UNAVAILABLE = "service_unavailable"
-    DEPLOYMENT_ISSUE = "deployment_issue"
-    RESOURCE_EXHAUSTION = "resource_exhaustion"
+    HIGH_ERROR_RATE = "high_error_rate"
+    INCREASED_LATENCY = "increased_latency"
+    RESOURCE_SPIKE = "resource_spike"
+    VERSION_REGRESSION = "version_regression"
+    STATISTICAL_OUTLIER = "statistical_outlier"
 
 
 class SeverityLevel(Enum):
@@ -19,6 +18,17 @@ class SeverityLevel(Enum):
     MEDIUM = "medium"
     HIGH = "high"
     CRITICAL = "critical"
+
+
+class AnomalyThresholds(BaseModel):
+    """Configuration for anomaly detection thresholds"""
+
+    error_rate_absolute: float = 60.0  # Absolute error rate threshold (%)
+    error_rate_increase: float = 20.0  # Percentage increase threshold
+    latency_increase: float = 50.0  # Percentage increase threshold
+    resource_increase: float = 40.0  # Percentage increase threshold
+    isolation_forest_contamination: float = 0.1
+    historical_window_hours: int = 24
 
 
 class AnomalyAlert(BaseModel):
@@ -35,3 +45,18 @@ class AnomalyAlert(BaseModel):
     context_data: Dict[str, Any]
     affected_metrics: List[str]
     recommended_actions: List[str]
+
+
+class AnomalyDetection(BaseModel):
+    """Represents a detected anomaly"""
+
+    anomaly_type: AnomalyType
+    severity: SeverityLevel
+    service_name: str
+    service_version: str
+    timestamp: datetime
+    current_value: float
+    baseline_value: Optional[float] = None
+    percentage_change: Optional[float] = None
+    description: str = ""
+    metadata: Dict[str, Any] = {}
